foundry/requirements.txt
# Azure AI SDK packages for Foundry 2.0 + MCP integration
azure-ai-projects>=1.0.0b8
azure-ai-agents>=1.2.0b6
azure-identity>=1.15.0
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
